+++ opencog/cython/CMakeLists.txt
@@ -21,7 +21,6 @@ INCLUDE_DIRECTORIES(
 # https://github.com/OpenKinect/libfreenect/blob/master/wrappers/python/CMakeLists.txt
 
 
-
 # direct file and path creation for stuff that doesn't need to be compiled
 file(MAKE_DIRECTORY opencog)
 # module init
@@ -38,25 +37,6 @@ TARGET_LINK_LIBRARIES(PythonEval
 	${PYTHON_LIBRARIES}
 )
 
-
-# Python module to be loaded by CogServer
-ADD_LIBRARY (PythonModule SHARED
-	PythonModule.cc 
-	PyMindAgent.cc
-	PyRequest.cc
-)
-
-TARGET_LINK_LIBRARIES(PythonModule
-	PythonEval
-	server
-	${PYTHON_LIBRARIES}
-)
-ADD_DEPENDENCIES(PythonModule agent_finder)
-
-IF (WIN32)
-	INSTALL (TARGETS PythonModule DESTINATION "lib${LIB_DIR_SUFFIX}/opencog")
-ENDIF (WIN32)
 INSTALL (FILES 
 	opencog/__init__.py
 	DESTINATION "${DATADIR}/python/opencog")
-
