+++ opencog/modules/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 IF (HAVE_SERVER)
 	ADD_SUBDIRECTORY(benchmark)
+	ADD_SUBDIRECTORY(python)
 	IF (HAVE_ZMQ AND TBB_FOUND)
 		ADD_SUBDIRECTORY(events)
 	ENDIF (HAVE_ZMQ AND TBB_FOUND)
