+++ opencog/modules/python/PyMindAgent.cc
@@ -1,10 +1,10 @@
 
-#include "PyMindAgent.h"
+#include <opencog/modules/python/PyMindAgent.h>
 
 #include <opencog/server/CogServer.h>
 
-#include "opencog/agent_finder_types.h"
-#include "opencog/agent_finder_api.h"
+#include <opencog/cython/opencog/agent_finder_types.h>
+#include <opencog/cython/opencog/agent_finder_api.h>
 
 using namespace opencog;
